changed js files
