+++ js/data.js
@@ -1,0 +1,42 @@
+const DESCRIPTIONS = [
+  'Моё фото на море.',
+  'Я и мой кот.',
+  'Создаем вид бурной деятельности.',
+  'Прогулка по парку.',
+  'Вид из окна.',
+  'Пришло время релакся.',
+  'Моё занятие в свободное время. А чем занимаетесь вы?',
+  'Во время просмотра любимого сериала.',
+  'Моё фото на конкурсе красоты.'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Антон',
+  'Артём',
+  'Семен',
+  'Анна',
+  'Ирина',
+  'Ольга',
+  'Никита',
+  'Глеб',
+  'Людмила',
+  'Иван',
+  'Анастасия',
+  'Слава',
+  'Дарья'
+];
+
+export {
+  DESCRIPTIONS,
+  COMMENTS,
+  NAMES
+};
